eplb reconstruct
